Add Custom Field

+++ db/migrate/20110323171713_add_custom1check_to_products.rb
@@ -1,6 +1,6 @@
 class AddCustom1checkToProducts < ActiveRecord::Migration
   def self.up
-    add_column :products, :custom1check, :tinyint
+    add_column :products, :custom1check, :int
   end
 
   def self.down
